pyproject.toml
[project]
name = "text-cleaning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.1",
    "transformers>=4.38.2",
    "accelerate>=0.27.2",
    "tqdm>=4.67.1",
    "levenshtein>=0.27.1",
    "scikit-learn>=1.7.0",
    "nltk>=3.9.1",
    "rouge-score>=0.1.2",
    "protobuf>=6.31.1",
    "fire>=0.7.0",
    "datasets>=2.18.0",
    "google-genai>=1.20.0",
    "huggingface-hub>=0.32.4",
    "strenum>=0.4.15",
    "python-dotenv>=1.1.0",
    "evaluate>=0.4.4",
    "jiwer>=4.0.0",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.13",
]

[tool.ruff]
line-length = 120

[tool.setuptools]
include-package-data = true
packages = ["text_cleaning"]
package-dir = {"" = "src"}

[project.scripts]
text-cleaning = "text_cleaning.main:main"
